+++ model/index.model.js
@@ -1,6 +1,6 @@
 const sequelize = require('../config/databaseSequelize');
 const User = require('./user.model');
-const Blog = require('./post.model');
+const Blog = require('./blog.model');
 
 User.hasMany(Blog, { foreignKey: { name: 'username', allowNull: false } });
 Blog.belongsTo(User, { foreignKey: { name: 'username', allowNull: false} });
